Correct problems found by valgrind.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gui/trunk@36188 72102866-910b-0410-8b05-ffd578937521

+++ Source/NSWindow.m
@@ -5379,7 +5379,7 @@ current key view.<br />
       aSize = [aDecoder decodeSize];
       [self setMaxSize: aSize];
 
-      [aDecoder decodeValueOfObjCType: @encode(int)
+      [aDecoder decodeValueOfObjCType: @encode(NSInteger)
                                    at: &level];
       [self setLevel: level];
